backend-orchestrator/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
curl_cffi==0.13.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
filetype==1.2.0
frozendict==2.4.7
google-ai-generativelanguage==0.10.0
google-api-core==2.30.0
google-auth==2.49.0
googleapis-common-protos==1.73.0
groq==0.37.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
multitasking==0.0.12
numpy==2.0.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
peewee==4.0.1
platformdirs==4.4.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.0
starlette==0.49.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
websockets==15.0.1
xxhash==3.6.0
yfinance==1.2.0
zstandard==0.25.0

QuantLib==1.41
py_vollib==1.0.1
qdrant-client==1.7.2
sentence-transformers==2.2.2
slowapi==0.1.9
pyjwt==2.12.1
supabase==2.28.2
